requirements.txt
cryptography==42.0.5
python-dotenv==1.0.1
pytest==8.0.0
matplotlib==3.8.4
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
requests==2.32.3